Pickups and sword Skill
I added the coin, tagged it and made it dissapear, I also have a counter in the pickup script incase it's needed to link to other scripts for additional life purposes. I added a Health Pick up item that dissapears when picked up for now. I added a sword and the animation for the sword skill. Currently if you have the sword and press the left mouse button it will play the sword attack animation.